refactor(api): split app.py into route submodules
Extract 2491-line app.py into focused route modules under api/routes/:
- health.py (23 lines) — health endpoint
- system.py (228 lines) — OpenAPI + internal/external probes
- memory.py (156 lines) — memory CRUD
- memes.py (222 lines) — meme management (9 handlers)
- cognitive.py (86 lines) — cognitive events & profiles
- chat.py (359 lines) — WebUI chat with SSE streaming
- tools.py (462 lines) — tool invoke with async callbacks
- naga.py (897 lines) — Naga bind/send/unbind + moderation

Infrastructure:
- _context.py: RuntimeAPIContext dataclass (shared import root)
- _naga_state.py: NagaState class (request dedup + inflight tracking)

app.py reduced to 333 lines: server class + thin delegation wrappers
that preserve test compatibility (no test API changes needed).

Updated 4 test files to monkeypatch route-module-level symbols
instead of app-module-level symbols.
