+++ src/Undefined/api/__init__.py
@@ -1,5 +1,6 @@
 """Runtime API server for Undefined main process."""
 
-from .app import RuntimeAPIContext, RuntimeAPIServer
+from ._context import RuntimeAPIContext
+from .app import RuntimeAPIServer
 
 __all__ = ["RuntimeAPIContext", "RuntimeAPIServer"]
